test: run repository-mutating tests against per-test isolated repositories
Tests that mutate whole-repository state (locales, push, sync, status)
previously ran serially against the shared repository. They now create a
throwaway repository per test via an it.scoped repo override, letting the
whole suite run concurrently. Tests that do not touch repository state,
like --help and error-path tests, keep the shared repository.

The isolatedRepo fixture takes no other fixtures as dependencies because
vitest resolves it.scoped override dependencies only against other scoped
fixtures, silently passing undefined for anything from test.extend.

Also flattens the vitest config to a single test config now that the
serial project is gone. This enabled the previously inert root-level
typecheck option and surfaced that concurrent tests must use
onTestFinished from the test context rather than the global import.
