+++ test/locale-remove.test.ts
@@ -1,0 +1,26 @@
+import { describe } from "vitest";
+
+import { isolatedRepo, it } from "./it";
+import { upsertLocale, getLocales } from "./prismic";
+
+it("supports --help", async ({ expect, prismic }) => {
+	const { stdout, exitCode } = await prismic("locale", ["remove", "--help"]);
+	expect(exitCode).toBe(0);
+	expect(stdout).toContain("prismic locale remove <code> [options]");
+});
+
+describe("with an isolated repository", () => {
+	it.scoped({ repo: isolatedRepo });
+
+	it("removes a locale", async ({ expect, prismic, repo, token, host }) => {
+		await upsertLocale("fr-fr", { repo, token, host });
+
+		const { stdout, exitCode } = await prismic("locale", ["remove", "fr-fr"]);
+		expect(exitCode).toBe(0);
+		expect(stdout).toContain("Locale removed: fr-fr");
+
+		const locales = await getLocales({ repo, token, host });
+		const locale = locales.find((l) => l.id === "fr-fr");
+		expect(locale).toBeUndefined();
+	});
+});
