+++ test/locale.test.ts
