+++ vitest.config.ts
@@ -5,29 +5,10 @@ export default defineConfig({
 		globalSetup: ["./test/setup.global.ts"],
 		forceRerunTriggers: ["**/dist/index.mjs"],
 		typecheck: { enabled: true },
-		projects: [
-			{
-				test: {
-					name: "concurrent",
-					setupFiles: ["./test/setup.ts"],
-					include: ["./test/**/*.test.ts"],
-					exclude: ["./test/*.serial.test.ts"],
-					sequence: { concurrent: true },
-					testTimeout: 30_000,
-					retry: 2,
-				},
-			},
-			{
-				test: {
-					name: "serial",
-					setupFiles: ["./test/setup.ts"],
-					include: ["./test/*.serial.test.ts"],
-					sequence: { concurrent: false },
-					fileParallelism: false,
-					testTimeout: 30_000,
-					retry: 2,
-				},
-			},
-		],
+		setupFiles: ["./test/setup.ts"],
+		include: ["./test/**/*.test.ts"],
+		sequence: { concurrent: true },
+		testTimeout: 30_000,
+		retry: 2,
 	},
 });
